+++ src/content.config.ts
@@ -1,14 +1,15 @@
 // 1. Import utilities from `astro:content`
-import { defineCollection, z } from 'astro:content';
+import { defineCollection } from "astro:content";
 
 // 2. Import loader(s)
-import { glob, file } from 'astro/loaders';
+import { glob } from "astro/loaders";
 
 // 3. Define your collection(s)
 const blog = defineCollection({
-    loader: glob({
-         pattern: "**/*.{md,mdx}", base: "./content/"
-    }),
- });
+  loader: glob({
+    pattern: "**/*.{md,mdx}",
+    base: "./content/",
+  }),
+});
 // 4. Export a single `collections` object to register your collection(s)
 export const collections = { blog };