+++ src/utils/date.ts
@@ -1,6 +1,5 @@
-import dayjs from 'dayjs';
-
+import dayjs from "dayjs";
 
 export function formatCreatedTime(createdTime: string) {
-    return dayjs(createdTime).format('YYYY-MM-DD')
+  return dayjs(createdTime).format("YYYY-MM-DD");
 }